feat(widgets): recent saves widgets, minor views refactor
